Fix issue with threading and add skeleton for Email plugin
Change the run method in the WorkerInterface to a slot so that it
can be called when the thread is started. This fixes a bug where
new plugins were not being run on the new threads. Also adds the
skeleton for an email plugin. This is largely untested but does
connect to the rabbit MQ server and attempts to send an email
using the QTcpSocket. Can't finish testing until a mail server is
set up.