Fix models and fix Smtp
Fix models by returning an empty object with ID set to -1 when
it is not found in the DB. Fix Smtp to create a perisitent
Smtp connection over which it can send multiple emails. Stops the
emails from being overwritten by newer ones. Untested with a working
SMTP mail server. Probably closes #3