Add email for banned logins
Adds the template file for the email for banned users trying to log
in. Also adds the email request proto and the banned user proto as
well as the model generator for generating it. Also adds the email
generator for generating banned login emails and registers it with
the email plugin.