Update project calculate task deadlines plugin
Use the graph builder to build a representation of the graph.
Use this to calculate the deadlines of each task for each sub graph.
Add the abilty to provide valid deadlines for graph that must run
shorter than the default estimate. Decrease the deadline period of each
task type at different times to ensure the longer tasks get more
time than others. #19